fix: #906 lab/flow/types/profiles/hardware review nit-batch
Behavior/correctness fixes:
- coding_task: reject setupContent keys that normalize to no file component
  (all-CurDir '.'/'./') — was passing validation then erroring opaquely on write
- profiles/swap: compare loaded context in u64 (was truncating LoadedModel.context
  to u32 before the sufficiency check → possible needless reload)
- lab/inspect: summary_chars is now a char count, not a byte count
- crew/index: pre-validate escalation hand-off targets resolve to a known role
  before insert — a dangling HandOffTo target now fails with a role-named message
  instead of an opaque deferred-FK abort that rolls back the whole rebuild
- crew/loader: 1 MiB size cap on manifest reads (defense-in-depth vs hostile/corrupt)
- crew/pins: embedded-asset invariants bail! instead of panic! (fn returns Result)
- doctor: role-pin-drift + agent-role-scaffold checks treat a TOCTOU NotFound
  (deleted between exists() and read) as Pass, not a spurious Warn
- hardware: Linux read_physical_cores counts unique (physical id, core id) pairs,
  falling back to logical count — the field is named physical_cores
- fleet/routing + roster: warn on a pre-epoch clock instead of silently stamping 0

Doc/INFO:
- lab/fixture_cli + coding_task: surface that verify_command runs on the HOST shell
  with operator privileges (warn at register; SECURITY note on run_verify_command)
- flow/integrity: document records_checked (record count, header-excluded) vs
  break_at_line (1-indexed file line) so cron tooling doesn't misread them
- coding_task: note the best-effort sandbox canonicalize fallback
- m_series_64: doc-drift — Small (0-32GB) tier is owned by m_series_32, not generic

+++ crates/darkmux-profiles/src/swap.rs
@@ -77,8 +77,11 @@ fn should_patch_openclaw(opts: &SwapOpts) -> bool {
 /// load just to shrink it — the operator who loaded it bigger has the RAM for
 /// it (operator sovereignty over loaded state). Only an *insufficient* load
 /// (smaller than the minimum) triggers a reload.
-fn ctx_sufficient(loaded_ctx: u32, wanted_n_ctx: u32) -> bool {
-    loaded_ctx >= wanted_n_ctx
+fn ctx_sufficient(loaded_ctx: u64, wanted_n_ctx: u32) -> bool {
+    // (#906) Compare in u64 — `LoadedModel.context` is u64; truncating it to
+    // u32 before the check could (in principle) wrap a very large loaded
+    // context down below the wanted minimum and trigger a needless reload.
+    loaded_ctx >= u64::from(wanted_n_ctx)
 }
 
 /// (#590) The context window at which swap loads the machine's standing
@@ -186,7 +189,7 @@ pub fn swap(profile: &Profile, registry: &ProfileRegistry, opts: SwapOpts) -> Re
             continue;
         }
         let desired_ctx = want.get(&cur.identifier).copied();
-        if desired_ctx.is_some_and(|d| ctx_sufficient(cur.context as u32, d)) {
+        if desired_ctx.is_some_and(|d| ctx_sufficient(cur.context, d)) {
             continue; // already loaded with enough context
         }
         if !opts.quiet {
@@ -207,7 +210,7 @@ pub fn swap(profile: &Profile, registry: &ProfileRegistry, opts: SwapOpts) -> Re
         .collect();
     for d in &desired {
         if let Some(c) = loaded_after_unload.get(d.identifier.as_str()) {
-            if ctx_sufficient(c.context as u32, d.n_ctx) {
+            if ctx_sufficient(c.context, d.n_ctx) {
                 continue; // already loaded with enough context (n_ctx is a min)
             }
         }
